+++ packages/restaflib/src/computeFetchData.js
@@ -5,9 +5,10 @@
 
  /**
  * 
- * @decription Fetch data from a SAS Table
+ * @description Fetch data from a SAS Table
  * @async
  * @module computeFetchData
+ * @category restaflib/compute
  * @param {object} store - restaf store
  * @param {computeSummary} computeSummary - Summary object created by computeSummary method
  * @param {string} table - name of the table